support get DOT chain data
